# Conflicts:
#	samples/ControlCatalog/MainView.xaml
#	src/Browser/Avalonia.Browser/BrowserPlatformSettings.cs
#	src/Windows/Avalonia.Win32/Win32Platform.cs
#	src/Windows/Avalonia.Win32/Win32PlatformSettings.cs

+++ src/Avalonia.Base/Media/GlyphMetrics.cs
@@ -10,15 +10,15 @@ public readonly record struct GlyphMetrics
     /// <summary>
     /// Distance from the top extremum of the glyph to the y-origin.
     /// </summary>
-    public int YBearing{ get; init; }
+    public int YBearing { get; init; }
 
     /// <summary>
     /// Distance from the left extremum of the glyph to the right extremum.
     /// </summary>
-    public int Width{ get; init; }
+    public ushort Width { get; init; }
 
     /// <summary>
     /// Distance from the top extremum of the glyph to the bottom extremum.
     /// </summary>
-    public int Height{ get; init; }
+    public ushort Height { get; init; }
 }
